Refactor bcstoj. Touchup jtobcs
